+++ lib/claide.rb
@@ -8,7 +8,7 @@ module CLAide
   #
   #   CLAide’s version, following [semver](http://semver.org).
   #
-  VERSION = '0.8.1'
+  VERSION = '0.8.2'
 
   require 'claide/ansi'
   require 'claide/argument'